pyproject.toml
[tool.poetry]
name = "three-project"
version = "0.1.0"
description = ""
authors = ["kiuter1 <ytigraman875@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
asyncmy = "^0.2.9"
alembic = "1.13.1"
pyjwt = {extras = ["crypto"], version = "^2.10.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.17"
pydantic = {extras = ["email"], version = "^2.10.3"}
miniopy-async = "^1.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
